Add typescript to jest. Fix all tests
