+++ babel.config.js
@@ -1,6 +1,7 @@
 module.exports = {
 	presets: [
-		"@babel/preset-env",
 		["@babel/preset-react", { runtime: "automatic" }],
+		"@babel/preset-env",
+		"@babel/preset-typescript",
 	],
 };
