+++ package.json
@@ -15,6 +15,7 @@
 	"devDependencies": {
 		"@babel/preset-env": "^7.19.0",
 		"@babel/preset-react": "^7.18.6",
+		"@babel/preset-typescript": "^7.18.6",
 		"@testing-library/jest-dom": "^5.16.2",
 		"@testing-library/react": "^13.4.0",
 		"@testing-library/user-event": "^14.4.3",
@@ -34,6 +35,7 @@
 		"obsidian": "^0.16.0",
 		"react-test-renderer": "^18.2.0",
 		"renamer": "^4.0.0",
+		"ts-jest": "^28.0.8",
 		"tslib": "^2.4.0",
 		"typescript": "^4.8.2"
 	},
