+++ src/services/random/index.ts
@@ -1,4 +1,4 @@
-import { COLOR } from "src/constants";
+import { COLOR } from "../../constants";
 
 export const randomColor = (): string => {
 	const index = Math.floor(Math.random() * Object.keys(COLOR).length);
