+++ src/services/string/modify/index.test.ts
@@ -1,4 +1,4 @@
-import { uppercaseFirst } from "./";
+import { uppercaseFirst } from ".";
 
 describe("uppercaseFirst", () => {
 	it("uppercases first letter", () => {
